+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Fixes
 - Choices workflow that either block the story or advance 2 times the story
+- Impossible for a character to talk in the story after an camera angle controller is stopped